+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 [Unreleased]
 ------------
 
+[3.6.0] (2020-12-15)
+--------------------
+
 ### Added
 
  - Provide base fragment controllers for content elements, modules and hybrids
@@ -219,6 +222,8 @@ Changelog
  - Added changelog (#9)
 
 [Unreleased]: https://github.com/netzmacht/contao-toolkit/compare/3.4.5...master
+[3.6.0]: https://github.com/netzmacht/contao-toolkit/compare/3.5.0...3.6.0
+[3.5.0]: https://github.com/netzmacht/contao-toolkit/compare/3.4.5...3.5.0
 [3.4.5]: https://github.com/netzmacht/contao-toolkit/compare/3.4.4...3.4.5
 [3.4.4]: https://github.com/netzmacht/contao-toolkit/compare/3.4.3...3.4.4
 [3.4.3]: https://github.com/netzmacht/contao-toolkit/compare/3.4.2...3.4.3